+++ ML-Frameworks/tensorflow-aarch64/CHANGELOG.md
@@ -8,7 +8,18 @@ where `YY` is the year, and `MM` the month of the increment.
 ## [unreleased]
 
 ### Added
- - Support for authenticated GitHub access to apply-github-patch
+
+### Changed
+
+### Removed
+
+### Fixed
+
+#### [r25.04] 2025-05-18
+https://github.com/ARM-software/Tool-Solutions/tree/r25.05
+
+### Added
+ - Support for authenticated GitHub access in apply-github-patch
 
 ### Changed